+++ CHANGELOG.md
@@ -11,15 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SHVPath` and its usage in `SHV*` classes
 - `SHVBase` now has `user_id` property that is default user ID used with `call`
   method
+- `RpcRequestInvalidError` and `RpcTryAgainLaterError`
 - `RpcTypeOptional` as easier alternative to `RpcTypeOneOf` when combining only
   with `rpctype_null`.
 
 ### Changed
 - Updated `getLogR` type hint according to spefication, it is now Struct
   instead of KeyStruct.
-
-
-### Changed
 - Websockets RPC Transport now requires `shv3` to be used as subprotocol.
   Connections failing to select this subprotocol won't be established.